Fixed additional QA issues and bugs found while testing
• Added a unit test to validate the GroupMember sorting continues to work as expected
• Updated the AppSetup process to be simpler (no need to check if it had previously run anymore)
• Removed some state management code from the NotificationServiceExtension (no longer needed now that state is properly managed via the Dependencies)
• Fixed an issue where if you had updated another client, gotten updated groups in your config, and then update the iOS client then it wouldn't create the updated groups until a UserGroups config change occurred
• Fixed a bug where we would incorrectly try to retrieve the disappearing messages settings for V2 Groups from the UserGroups config instead of the GroupInfo one
• Fixed an issue where the updated groups poller might not get started correctly in some cases
• Fixed an issue where we could incorrectly add a "you were invited..." control message on linked devices when creating an updated group
• Fixed an issue where the "open url" modal wouldn't be dismissed when copying the url
• Fixed an issue where reactions could appear on locally deleted community messages
• Fixed an issue where the "Note to Self" conversation could be mislabelled in the share extension
• Fixed an issue where sharing a url with a preview would fail
• Fixed an issue where a quote for an attachment wouldn't show the thumbnail if the conversation was open when the quote message was received
• Fixed an issue where the background colour of the display picture could be incorrect when in a multi-avatar for a group conversation

+++ SessionMessagingKit/Sending & Receiving/MessageReceiver.swift
@@ -543,22 +543,28 @@ public enum MessageReceiver {
         openGroupMessageServerId: Int64,
         openGroupReactions: [Reaction]
     ) throws {
-        guard let interactionId: Int64 = try? Interaction
-            .select(.id)
-            .filter(Interaction.Columns.threadId == threadId)
-            .filter(Interaction.Columns.openGroupServerMessageId == openGroupMessageServerId)
-            .asRequest(of: Int64.self)
-            .fetchOne(db)
-        else {
-            throw MessageReceiverError.invalidMessage
+        struct Info: Decodable, FetchableRecord {
+            let id: Int64
+            let variant: Interaction.Variant
         }
         
+        guard let interactionInfo: Info = try? Interaction
+            .select(.id, .variant)
+            .filter(Interaction.Columns.threadId == threadId)
+            .filter(Interaction.Columns.openGroupServerMessageId == openGroupMessageServerId)
+            .asRequest(of: Info.self)
+            .fetchOne(db)
+        else { throw MessageReceiverError.invalidMessage }
+        
+        // If the user locally deleted the message then we don't want to process reactions for it
+        guard !interactionInfo.variant.isDeletedMessage else { return }
+        
         _ = try Reaction
-            .filter(Reaction.Columns.interactionId == interactionId)
+            .filter(Reaction.Columns.interactionId == interactionInfo.id)
             .deleteAll(db)
         
         for reaction in openGroupReactions {
-            try reaction.with(interactionId: interactionId).insert(db)
+            try reaction.with(interactionId: interactionInfo.id).insert(db)
         }
     }
